[MOD/#181] CalendarState 로직 및 이름 수정
