+++ feature/src/main/java/com/terning/feature/calendar/list/CalendarListScreen.kt
@@ -38,7 +38,7 @@ import com.terning.feature.R
 import com.terning.feature.calendar.calendar.CalendarViewModel
 import com.terning.feature.calendar.calendar.component.CalendarDialog
 import com.terning.feature.calendar.calendar.model.CalendarDefaults.flingBehavior
-import com.terning.feature.calendar.calendar.model.CalendarState.Companion.getDateByPage
+import com.terning.feature.calendar.calendar.model.CalendarModel.Companion.getDateByPage
 import com.terning.feature.calendar.list.component.CalendarScrapList
 import kotlinx.coroutines.flow.distinctUntilChanged
 import java.time.LocalDate
