+++ feature/src/main/java/com/terning/feature/calendar/month/component/HorizontalCalendar.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import com.terning.core.designsystem.theme.White
 import com.terning.domain.entity.response.CalendarScrapModel
 import com.terning.feature.calendar.calendar.model.CalendarDefaults.flingBehavior
-import com.terning.feature.calendar.calendar.model.CalendarState.Companion.getDateByPage
+import com.terning.feature.calendar.calendar.model.CalendarModel.Companion.getDateByPage
 import com.terning.feature.calendar.month.model.MonthData
 import java.time.LocalDate
 import java.time.YearMonth
